% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_rn.R
\docType{data}
\name{dat_rn}
\alias{dat_rn}
\title{Example data}
\format{
A data frame with 10 rows and 10 variables:
\describe{
  \item{X1}{Litter ID}
  \item{X2}{Dose of ethylene glycol (EG) administered}
  \item{X3}{Fetal weight}
  \item{X4}{Indicator for fetal malformation}
}
}
\source{
http://users.stat.ufl.edu/~aa/
}
\usage{
dat_rn
}
\description{
Data from rnorm
}
\references{
Agresti, Alan.
Categorical data analysis. Vol. 482.
John Wiley & Sons, 2003.Harvard
}
\keyword{datasets}
