% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhf1.R
\name{max_hhz}
\alias{max_hhz}
\title{Max}
\usage{
max_hhz(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
y
}
\description{
a function that returns maximum value
}
\examples{
max_hhz(c(1,4,NA,10,-500))
}
\author{
Lehang Zhong
}
